Multiple
Render the datetime correctly. It is currently showing as "ISO standard: diagnostic validation (Project) - 2021-01-08 HH:00" (0.1 - 0.3)
The last project priority is not being displayed (0 - 0.5) - this was caused by the bug rendering the datetime as "HH:00"
Latest priority date is "2020-12-16", this should be changed to include today's date if it is later than last date. (0.1)
Project priority events in 2021 not showing due to some bug in sort of datalist ticks not being natural